requirements.txt
# Core dependencies
feedparser>=6.0.10
requests>=2.31.0
PyYAML>=6.0.1

# Web scraping dependencies
beautifulsoup4>=4.14.0
trafilatura>=2.0.0

# LLM providers
google-generativeai>=0.3.0
openai>=1.0.0

# Template rendering
jinja2>=3.1.2

# Development and testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality and linting
ruff>=0.1.0
mypy>=1.5.0

# Type stubs
types-requests>=2.31.0
types-PyYAML>=6.0.0